requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
idna==2.10
python-dotenv==0.15.0
redis==3.5.3
requests==2.25.0
rq==1.7.0
urllib3==1.26.2
